First commit. Version 2 of financial planning tool
This marks the beginning of version 2 of the financial planning tool.

The focus of this version is on improving accessibility, architecture, and overall maintainability. To that end, the backend document organization has been restructured, and the codebase will undergo optimization in terms of performance and visual design.

At this point, only the folder structure has been set up, and minimal placeholder files have been created to support development using FastAPI (for the backend) and Streamlit (for the frontend).